move scan tests into tests/lax_control_flow_test.py
